♻️refactor : 멤버 폴더명 변경
